Change categoryDetail middleware
First route handler calls the findById method on the Category model to
get information for a specific ID, find method on the Product to get
all products that have same associated category ID.

If the category does not exist in the database then we render a
"not found" page.

if it is exist then we passed results to render category detail
template.

and sorting the results by the name using the sort method and then
render specify the categoryList template and passing the results.