Change productDetail middleware
The route handler calls the findById method on the Product model to
get information for a specific ID and also calls populate method on
Product, specifying the category field to get the category detail.

If the product does not exist in the database then handler will render
a "not found" page.

if the product is exist then handler will pass results to render
category detail template.