Change CheckPrefersColorScheme method
Method will be renamed to getDarkScheme and retrieves the value of
the local storage "darkScheme" key.

If the value does not exist, return null, then call the setDarkScheme
method to get the result, and then set whether to add a "dark" class
name to the container block based on the true or false value of the
result.

If true, sets the "dark" class name to the container block.