Change categoryCreatePost function
The controller is an array of middleware function.

First we validate and sanitize the data.

If the data is valid, call the isCategoryExist method to check whether
the category entered by the user exists. If it exists, redirect to its
URL. If it does not exist, it will be created and redirected to the
new category URL.

If the data is invalid, call the renderErrorMessages method to
redisplay the form page with error message and the data that was
originally input by the user.