Change productCreatePost function
First we validate and sanitize all text type data and validate a file
type date.

If the data is valid, call the "isProductExist" function to check
whether the product name entered by the user exists. If it exists,
redirect to its details page. If it does not exist, it will be called
"createProduct" executes the "uploadFile" and "addNewProduct" functions.

If the data is invalid, the "renderErrorMessages" function is called to
redisplay the product form page, which contains the error message and
the data originally entered by the user.

"uploadFile" checks whether the width and height of the image file are
greater than 800, and If so, resizes it to 800 and uploads it to google
cloud storage.

"addNewProduct" saves data to MongoDB.