Change categoryUpdatePost function
Rename variable existingCategory to category

Use "try" and "catch" to handle errors.

In "try" we use a condition to check if the category
is found, if not then call "next" and pass the result of the
"createError" function, if so then check if NODE_ENV environment
variable is "production" and category variable has not a
"expiresAfter" if so then we will redirect to the category's
details page, if not, we will call the "validationFields" function
to check all fields.

In "catch" we call "next" and pass the result of the
"createError" function.