Change categoryDeletePost function
If the category is found, then check if NODE_ENV environment
variable is "development" or category variable has a
"expiresAfter" key, if so then calls "deleteCategory" function
to render "categoryDetail" template, if not, we will redirect to
the category's details page.

The "deleteCategory" function will delete category with the
specified id and then redirect to categories page.