Change productCreatePost function
The find method only gets the name field on the category model.

In each field, "errorMessage" property will be put into the specific
validator.

Add a custom validator for the name field to verify that the value
exists in the database.

The product variable is now an object instead of product model
instance.

Replace "isProductExist" to "createProduct"