Change productForm template
Use a condition In class attribute of div element with class
image, if product is truly, set class to 'preview', if not,
set for empty string.

Use a condition In src attribute of img element, if product
is truly, set src to product's image url, if not, set for
empty string.