Change productUpdateGet function
Use "try" and "catch" to handle errors.

In "try" we use a condition to check if the "currentProduct"
is found, if not then call "next" and pass the result of the
"createError" function, if so then check whether NODE_ENV environment
variable is "development" or "currentProduct" has a "expiresAfter"
property, if so then we calls "validationFields" function, if not,
we will redirect to the "currentProduct's" details page.

The "validationFields" function validates all fields. If there
is an error in the result, "renderErrorMessages" will be called
to render the productForm template with the error messages.
If not, "updateProduct" will be called to handle the update.

In "catch" we call "next" and pass the result of the "createError"
function.