Change productDeleteGet function
Use "try" and "catch" to handle errors.

In "try" we use a condition to check if the "product"
is found, if not then call "next" and pass the result of the
"createError" function, if so then check whether NODE_ENV environment
variable is "development" or "product" has a "expiresAfter"
property, if so then we renders "productDetail" template, if not,
we will redirect to the "product's" details page.

In "catch" we call "next" and pass the result of the "createError"
function.