Change productDeletePost function
Use "try" and "catch" to handle errors.

In "try" we use a condition to check if the "product"
is found, if not then call "next" and pass the result of the
"createError" function, if so then check whether NODE_ENV environment
variable is "development" or "product" has a "expiresAfter"
property, if so then we calls "handleDelete" function, if not,
we will redirect to the "product's" details page.

The handleDelete function will call "deleteProductImage" to delete
the image stored in the storage, and "deleteProduct" to delete the
data stored in the database.

In "catch" we call "next" and pass the result of the "createError"
function.