Change productUpdatePost function
In "validationFields" function, if all schemas were not errors
and upload image is not error, then calls updateProduct.

Add a condition to "uploadImageError" variable, If "uploadImage"
is true then calls "validateUploadImage" function.

Add imageUrl property to product variable.

In the "updateProduct" function, a "RenameProductImage" function
has been added. If the uploadImage variable is false,
"RenameProductImage" is called. If not, "handleUpdateImage"
is called.